fixup by @hackaugusto suggestions
